backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-anthropic
langchain-text-splitters
langchain-chroma
langchain-openai
python-dotenv
pypdf
ragas
